requirements.txt
python-dotenv
numpy
pandas
scikit-learn
tensorflow
pymongo
pymongo[srv]==3.6
certifi

# -e.